+++ src/main/java/yesman/epicfight/api/client/input/controller/IEpicFightControllerMod.java
@@ -1,6 +1,5 @@
 package yesman.epicfight.api.client.input.controller;
 
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import yesman.epicfight.api.client.input.InputMode;
 import yesman.epicfight.api.client.input.PlayerInputState;
@@ -21,7 +20,6 @@ import yesman.epicfight.api.client.input.action.EpicFightInputAction;
 /// **Note:** This interface exposes low-level controller integration.
 /// Most consumers should use a higher-level abstraction unless direct access is necessary for functionality
 /// that cannot be achieved otherwise.
-@ApiStatus.Experimental
 public interface IEpicFightControllerMod {
     /// Returns the controller mod’s display name (e.g., `Controlify`).
     /// Intended for logging or debugging only; should not influence gameplay logic or be used as a workaround.
@@ -30,19 +28,6 @@ public interface IEpicFightControllerMod {
     /// Returns the current input mode (keyboard/mouse or controller).
     @NotNull
     InputMode getInputMode();
-
-    /// Retrieves the universal controller binding for a given Epic Fight input action.
-    ///
-    /// This method maps actions such as [EpicFightInputAction#JUMP],
-    /// [EpicFightInputAction#ATTACK], and
-    /// [EpicFightInputAction#MOVE_FORWARD] to the corresponding controller buttons or axes.
-    ///
-    /// @param action the Epic Fight input action to retrieve the binding for.
-    /// @return the [ControllerBinding] associated with the specified action,
-    /// providing both state and metadata about the input.
-    /// @see ControllerBinding
-    @NotNull
-    ControllerBinding getBinding(EpicFightInputAction action);
 
     /// Retrieves the current input state.
     /// This is used internally by Epic Fight to perform actions such as [EpicFightInputAction#DODGE].
@@ -53,11 +38,8 @@ public interface IEpicFightControllerMod {
 
     /// Checks whether the specified input actions are bound to the same controller button.
     ///
-    /// This comparison only applies to digital buttons (e.g., A, B, L3, R3) and does not
-    /// account for analogue inputs such as triggers or stick movements.
-    ///
     /// @param action  the first input action
     /// @param action2 the second input action
     /// @return `true` if both actions are bound to the same controller button; `false` otherwise
-    boolean isBoundToSameButton(@NotNull EpicFightInputAction action, @NotNull EpicFightInputAction action2);
+    boolean isBoundToSamePhysicalInput(@NotNull EpicFightInputAction action, @NotNull EpicFightInputAction action2);
 }